Further cleanup in progress
